@@ -2,11 +2,11 @@ package org.icatproject.ids.exceptions;
 
 import java.io.ByteArrayOutputStream;
 
-import javax.json.Json;
-import javax.json.stream.JsonGenerator;
-import javax.ws.rs.core.Response;
-import javax.ws.rs.ext.ExceptionMapper;
-import javax.ws.rs.ext.Provider;
+import jakarta.json.Json;
+import jakarta.json.stream.JsonGenerator;
+import jakarta.ws.rs.core.Response;
+import jakarta.ws.rs.ext.ExceptionMapper;
+import jakarta.ws.rs.ext.Provider;
 
 /**
  * Capture any {@link org.icatproject.ids.exceptions.IdsException WebServiceException} thrown from
